// src/api/generated/models/auditRecordVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AppVO } from './appVO';

export interface AuditRecordVO {
  id?: string;
  appId?: string;
  app?: AppVO;
  status?: number;
  auditorId?: string;
  remark?: string;
  auditTime?: string;
  createdAt?: string;
}