// src/api/generated/models/pageResultAuditRecordVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AuditRecordVO } from './auditRecordVO';

export interface PageResultAuditRecordVO {
  list?: AuditRecordVO[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}